+++ 3rd/tts-server/android-yolo/app/src/test/java/com/aasc/yolo/CpuModeTest.kt
@@ -10,9 +10,20 @@ class CpuModeTest {
     }
 
     @Test
-    fun persistedValuesMapToThreeCpuModes() {
+    fun persistedValuesMapToFiveCpuModes() {
         assertEquals(CpuMode.AUTO, CpuMode.fromPersistedValue(0))
         assertEquals(CpuMode.BIG, CpuMode.fromPersistedValue(1))
         assertEquals(CpuMode.LITTLE, CpuMode.fromPersistedValue(2))
+        assertEquals(CpuMode.SINGLE_BIG, CpuMode.fromPersistedValue(3))
+        assertEquals(CpuMode.SINGLE_LITTLE, CpuMode.fromPersistedValue(4))
+    }
+
+    @Test
+    fun singleCoreModesUseOneOrtThreadAndClusterModesKeepTwo() {
+        assertEquals(2, CpuMode.AUTO.intraOpThreads)
+        assertEquals(2, CpuMode.BIG.intraOpThreads)
+        assertEquals(2, CpuMode.LITTLE.intraOpThreads)
+        assertEquals(1, CpuMode.SINGLE_BIG.intraOpThreads)
+        assertEquals(1, CpuMode.SINGLE_LITTLE.intraOpThreads)
     }
 }